feat: implement regex anchors `^` and `$` in multi-line mode.
Until now, anchors `^` and `$` only worked in a single-line mode, where `^`matches only at the start of the data and `$` at the end. In multi-line mode these anchors also match at the start and the end of a line respectively. Multi-line mode can be enabled by prefixing the regex with `(?m)`.