perf: memoized some module functions (#311)
There are a bunch of functions in the elf, pe, and macho modules who's results can be memoized (imphash, import_md5, etc). This can result in a significant performance boost in cases where a large number of calls to these functions are made (usually because a large number of rules are used in a single scanner).

Co-authored-by: Victor M. Alvarez <vmalvarez@virustotal.com>