refactor: don't rely on the `twistrs` crate for determining if domains are homoglyphs
The `vt` module implements a `permutation_of` method that determines returns true if a domain is a permutation of another domain. Homoglyphs is one of the supported permutations, but generating all possible homoglyphs with the `twistrs` crate is extremely slow, and time grows exponentially with the domain's length.

With this change we implement our own logic for this, which runs in linear time.