fix: improve the `text_as_hex` warning.
The `text_as_hex` warning is shown when an hex string is used where a literal text string could do the job. For instance, `{61 61 61 61}` can be rewritten as "aaaa", which is more legible. However, the warning was appearing in cases where the string contained non-printable ASCII characters. For instance, the warning was proposing to replace `{ 50 4B 03 04 }` with "PK\u{3}\u{4}".

This fix makes the warning more strict, and is only shown when the literal text strings doesn't contain non-printable characters, except for tabs, newlines and line breaks.