refactor: stop using the `twistrs` crate for typosquatting detection in the `vt` module.
The `twistrs` crate was not suitable for our needs for performance reasons. The `twistrs` was created with the goal of generating every possible typosquatting domain for a given domain. In our case we have two domains, the legitimate one, and some other domain that could be a potential typosquatting attempt for the legitimate one. By having both domains, we don't need to generate every typosquatting combination and some checks can be done a lot faster.