Add module invoke support to python bindings.
This commit adds the ability to dump module output from python bindings. The API
is roughly:

yara_x.Module('pe').invoke(data)

I chose to use an API that requires you instantiate a new Module object because
it will allow us to put other things as methods on it. I'm thinking we will want
to support things like invoke a module on a path or file object. This sets us up
to have a cleaner API in the future.

I also moved and renamed the SupportedModules enum out of the dump command so it
is now in a better place to be shared by the bindings and dump command.  This
allows us to raise a ValueError if you try to instantiate a new module with an
unsupported dump module.

As part of this I also realized that when using "dump" and outputting YAML (the
default) the keys would be different from outputting in JSON. For example, in
YAML a PDB path would be "pdb_path" while in JSON it was "pdbPath". To better
align these I made the JSON output use proto_field_true when serializing the
protobuf to JSON.