Do not use SupportedDumpModules in public API.
We are now using multiple enums for supported modules for "dumping" information.
One enum is in the dump CLI and the other is in the python bindings. While this
may be more confusing for developers it is a cleaner approach than exposing the
enum in the public API when it isn't ever used for anything public facing.