feat: add module invoke support to python bindings. (#348)
This commit adds the ability to dump module output from python bindings. The API
is roughly:

```
yara_x.Module('pe').invoke(data)
```

The API that requires that you instantiate a new `Module` object because
it will allow us to put other things as methods on it. This sets us up
to have a cleaner API in the future.

As part of this I also realized that when using "dump" and outputting YAML (the
default) the keys would be different from outputting in JSON. For example, in
YAML a PDB path would be "pdb_path" while in JSON it was "pdbPath". To better
align these I made the JSON output use proto_field_true when serializing the
protobuf to JSON.

---------

Co-authored-by: Victor M. Alvarez <vmalvarez@virustotal.com>