fix: CST to AST conversion aborted early due to max depth being reached.
The current depth was not being decremented when a rule had modifiers like `private` or `global`. When too many rules had any of those modifiers, the maximum depth was reached, and subsequent rules ignored. Now we make sure that the depth is zero after each source file is parsed, so that this kind of error is detected earlier next time.

Closes: #349.