feat: implement `include` statements. (#350)
This change introduces a new `include` AST node and updates the parser to recognize and process `include` keywords, enabling the inclusion of external YARA rule files. The `AST` now uses a vector of `Item` to store different types of top-level declarations, enabling to store imports, includes, and rules.

It also implements the `Compiler::add_include_dir` API to specify include directories and introduces a new compiler option to disable `include` statements. This is useful in scenarios where including external files poses a security risk or is simply not desired. When includes are disabled, the compiler will now produce an error if an `include` statement is encountered during compilation.