feat: warning for unsatisfiable expressions
Adds a warning that is emitted when a boolean expression can't be satisfied, for example, when comparing a lowercase string with a string containing uppercase characters. One common case is comparing the result of the hashing functions in the `hash` module (which return lowercase strings) with an upper case hash. These comparisons can't never succeed.