feat: implement constraints for TypeValue (#354)
Now a TypeValue can have additional constraints that gives the compiler information about the values. For instance, a TypeValue of type string can have a constraint that specifies that the value is always lowercase. These constraints allow to raise warnings in situations where a condition can't be met, like comparing a lowercase string with some other string that contains uppercase characters.