feat: add support for deprecated field warnings.
This commit introduces the ability to mark fields in protobuf definitions as deprecated. When a deprecated field is accessed in a YARA rule, a warning is issued, informing the user that the field is no longer recommended for use. The warning message is derived from the `deprecation_msg` option in the proto file.