fix: less strict scheme of cuckoo report (#369)
I noticed that in the original yara the scheme of cuckoo JSON report is less strict. For example report does not have to contain network attribute. In contrast with this behavior yara-x enforces a presence of some attributes, otherwise it does not match a sample, because of parsing error in serde.

Therefore I relaxed cuckoo JSON scheme to do not enforce a presence of any attributes (all of them are optional now).