fix: edge case with a rule that contains an invalid regexp and the rule name is reused
In such cases the compiler was panicking because the first rule  appeared in the symbol table even if it was not properly compiled because of the invalid regexp. When the second rule is found, the `check_for_existing_identifier` function panics because the identifier is found in the symbol table but the rule doesn't exist.

The fix consists in adding the symbol to the symbol table only when the rule is completely compiled and can't fail.