Fix tokenizer regex so it recognizes identifiers that start with underscore.
Prior to this change _123 would not be recognized as an identifier, but with
this change we now recognize it as an identifier because literals can no longer
start with an underscore.

Now literals must start with an number (or hex character if in base16) and can
use only a single underscore to separate groupings - double underscores are not
supported.

This is valid:

0xaa_bb

This is invalid:

0xaa__bb