refactor: more modular design for `Struct::from_proto_descriptor_and_msg`.
Previously, this function was the responsible of adding methods to structures. Now the methods are added only if necessary by a callback function that has the opportunity to modify the structure before it is finally returned.

This implies that methods are only added when they are actually needed. For instance, the scanner doesn't need structure with methods, only the compiler does.