feat: return an error if a regexp pattern has a prefix that can match an arbitrarily long sequence of arbitrary bytes.
For instance, the pattern `/.*foo/s` is rejected because of the `.*` prefix. This pattern is problematic as it will match at every offset in the file prior the occurrence of the `foo` string, which leads to a huge amount of matches. In the other hand, using a patterns that like doesn't make any sense at all.