feat(macho): implement symhash as defined by anomali (#425)
This change migrates the mach-o module's sym_hash which was implemented as a hash of all symbol table entries upon original design to the industry standard symhash defined by Anomali here:

http://github.com/threatstream/symhash
https://www.anomali.com/blog/symhash

This is a breaking change for mach-o rules, as `sym_hash()` no longer exists. I opted to rename it to `symhash()` to align with other tools and mach-o context.

This involved pulling out the n_type value from the Nlist parsed structs and exposing it via tags in the symbol table entries.