feat: raise warning for `0 of them` expressions.
`0 of them` is an ambiguous part of the language which should be avoided in favor of `none of them`. To grasp this, let’s consider the meaning of `2 of them`, which is `true` if two or more of the patterns match. Historically, `0 of them` followed this principle and would evaluate to true if at least one of the patterns matched. This ambiguity was resolved in YARA 4.3.0 by making `0 of them` evaluate to `true` if exactly zero patterns match. YARA-X follows the same logic, but to enhance clarity and avoid confusion, it’s recommended to use `none  of them`.