fix: panic when compiling certain rules with file size bounds.
This is an edge case that occurs when a rule A defines a pattern P, but rule A fails to compile due to an error in some other pattern defined later. If another rule B defines exactly the same pattern P, and both rules A and B had different file size bounds, a panic occurs.

Closes #481