fix: check address of functions for 0x0 (#482)
Added condition of parsing PE exports. Exports should not be parsed if AddressOfFunctions is 0x0, because:

Time/memory is wasted for nonse exports
Bad number of exports is provided (e. g. max number of PE exports)
This fix causes skip of the parsing of exports if AddressOfFunctions is 0x0 and therefore pe.number_of_exports is 0 in this case.