perf: avoid emitting code that invokes `search_for_patterns` if not necessary. (#487)
Until now, every time that a pattern was found in a condition, some WASM instructions were emitted to lazily invoke `search_for_patterns` if necessary. However this is not necessary in all cases, for instance:

```
$a and $b and $c
```

In the condition above, once the pattern `$a` has been processed, and the code that
invokes `search_for_patterns` has been emitted, we don't need to emit the same code
again, for `$b` and `$c`. The emitted code would actually checks for a global flag before invoking `search_for_patterns`, so this function is invoked once anyways, but we don't even need to emit the code that checks for the global flag (which takes time, bloats the WASM code, and increases compile time).

Now, the code that invokes `search_for_patterns` is emitted only when necessary.