fix: when the `--tag` argument is used, take into account while counting matching files.
The `--tag` arguments filters out the matching rules and reports only those that have the given tag. However, when counting the number of matching files the `--tag` filter was not taken into account, which resulted in the CLI reporting N matching files (because they where actually matching some rule), but the match not appearing in the output because the rule didn't have the requested tag.

With this fix the number of matching files reflects the number of files that actually appeared in the CLI output, excluding those that didn't appear because of the `--tag` filter.