fix: allow deprecated cargo_bin
81a4578aa03bd8980385d18fe355f05114c9cb36 updated assert_cmd to 2.1.1 and
assert_cmd::Command::cargo_bin was deprecated[1] in 2.1.0 so now rust-analyzer
is complaining about the usage. However, trying to move to the recommended
assert_cmd::cargo::cargo_bin_name!() doesn't work because that uses an
environment variable that is set at build time, and contains the path to the
binary being built. However, cli tests need to run the "yr" binary which doesn't
work with the recommended macro.

See https://docs.rs/assert_cmd/2.1.1/src/assert_cmd/macros.rs.html#85-92 for
details, but the TLDR is that cargo_bin_cmd!() expects to find the path to the
binary to run at an environment variable named CARGO_BIN_EXE_yr and that is not
set. There may be some trickery to set it during the build but I haven't figured
it out yet, nor do I know how to make this work in a platform independent way.

For now I'm just marking them deprecated.

[1]: https://docs.rs/assert_cmd/2.1.1/assert_cmd/cmd/struct.Command.html#method.cargo_bin