Implement "deps" command.
Given a set of rules parse it and walk the AST to find identifiers and generate
a dot file of them that can be fed into graphviz for visualization.

By default it generates a graph of all the rules but you can select any number
of rules with the -r argument.

For example, given these rules:

```
rule a { condition: pe.is_dll() }
rule b { condition: a }
rule c { condition: b }
rule d { condition: false }
```

And selecting using `-r b` you get output that looks like this:

```
digraph {
  b [fillcolor=paleturquoise, style="filled"];
  a [fillcolor=paleturquoise, style="filled"];
  pe [fillcolor=palegreen, style="filled"];
  a -> pe;
  b -> a;
}
```

This mode is best thought of as "what is the minimum set of rules and imports I
need to execute the selected rule."

Using the -R argument displays the reverse dependencies of a rule. For the same
rules above the output when using -R is:

```
digraph {
  b [fillcolor=paleturquoise, style="filled"];
  c [fillcolor=paleturquoise, style="filled"];
  c -> b;
}
```

This mode is best thought of as "if I change this rule, what other rules do I
also need to test."