fix: issue when reusing patterns that were defined by a rule that failed.
When a rule defines a pattern for the first time, but the compilation of that rule fails, the compiler is left in an inconsistent state that will cause issues with other rules that re-use the pattern defined by the failed rule.

This happens because the patterns hash map in the compiler was not restored to its previous state when `restore_snapshot` is called.