feat: add `inventory` feature for using the `inventory` crate as a replacement for `linkme`. (#504)
The linkme crate is being used for maintaining a global registry of functions that are callable from WASM. This PR introduces an alternative by using the inventory crate instead of linkme. Two new features named inventory and linkme have been introduced, by default linkme is enabled, and inventory is disabled.

The linkme crate doesn't work in all platforms (namely, WASM and Windows with the GNU compiler), inventory can be used in such cases.