refactor: remove some fields.
I prefer being conservative with the fields that are added to a module. Adding a new field later on is always possible if there's demand for it, but removing a field is a backward-incompatible action. I honestly don't feel that the removed fields are actually useful for a rule, but I'm open to change my mind if I see some evidence. In the case of the `xxxxx_size` fields, an alternative is using the `.len()` for the corresponding array.