fix: parser issue that puts under an `ERROR` node code that is valid.
In some cases the parser produces a CST with `ERROR` nodes that contain portions of code that are perfectly valid. For instance:

```
rule test { condition: true }
foo
```

In the code above, the rule `test ` is valid, but it was being put under an `ERROR` node in the CST because of the trailing `foo`. This commit fixes that.